controlador areas comunes métodos agregados

+++ BackendResidentes/src/java/ENTIDADES/Areacomun.java
@@ -6,6 +6,7 @@
 package ENTIDADES;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 import java.math.BigInteger;
 import java.util.List;
 import javax.persistence.Basic;
@@ -57,7 +58,7 @@ public class Areacomun implements Serializable {
     @Basic(optional = false)
     @NotNull
     @Column(name = "Capacidad")
-    private BigInteger capacidad;
+    private BigDecimal capacidad;
     @Basic(optional = false)
     @NotNull
     @Size(min = 1, max = 255)
@@ -78,7 +79,7 @@ public class Areacomun implements Serializable {
         this.areacomunPK = areacomunPK;
     }
 
-    public Areacomun(AreacomunPK areacomunPK, String nombre, String tipo, BigInteger capacidad, String descripcion) {
+    public Areacomun(AreacomunPK areacomunPK, String nombre, String tipo, BigDecimal capacidad, String descripcion) {
         this.areacomunPK = areacomunPK;
         this.nombre = nombre;
         this.tipo = tipo;
@@ -86,7 +87,7 @@ public class Areacomun implements Serializable {
         this.descripcion = descripcion;
     }
 
-    public Areacomun(BigInteger idArea, BigInteger conjuntoIdConjunto) {
+    public Areacomun(BigDecimal idArea, BigDecimal conjuntoIdConjunto) {
         this.areacomunPK = new AreacomunPK(idArea, conjuntoIdConjunto);
     }
 
@@ -114,11 +115,11 @@ public class Areacomun implements Serializable {
         this.tipo = tipo;
     }
 
-    public BigInteger getCapacidad() {
+    public BigDecimal getCapacidad() {
         return capacidad;
     }
 
-    public void setCapacidad(BigInteger capacidad) {
+    public void setCapacidad(BigDecimal capacidad) {
         this.capacidad = capacidad;
     }
